requirements.txt
flask
flask-sqlalchemy
flask-login
bcrypt
pymysql
ollama
python-dotenv
